+++ plugin-maven/CHANGES.md
@@ -4,7 +4,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 
 ## [Unreleased]
 ### Added
-* CleanThat Java Refactorer ([#???](https://github.com/diffplug/spotless/pull/???))
+* CleanThat Java Refactorer ([#1560](https://github.com/diffplug/spotless/pull/1560))
 
 ## [2.32.0] - 2023-02-05
 ### Added